requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.1.0
cryptography==44.0.0
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
filelock==3.17.0
HTMLParser==0.0.2
importlib_resources==6.5.2
ini2toml==0.15
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
keyring==25.6.0
mod==0.3.0
pandas==2.2.3
path.py==12.5.0
Pillow==11.1.0
pkg1==0.0.3
praw==7.8.1
protobuf==5.29.3
pyOpenSSL==25.0.0
pytest==5.4.3
python-dotenv==1.0.1
redis==5.2.1
Sphinx==8.1.3
thread==2.0.5
tomli_w==1.2.0
trove_classifiers==2025.1.15.22
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
xx==3.3.2
